# server/helpers/swagger.yaml
basePath: /
consumes:
    - application/json
    - multipart/form-data
...
info:
    description: |-
        the purpose of this application is to provide an application
        that is using plain go code to define an API

        This should demonstrate all the possible comment annotations
        that are available to turn go code into a fully compliant swagger 2.0 spec
    license:
        name: Apache-2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.txt
    title: Meshplay API.
    version: 0.4.27
paths:
...
    /api/user/performance/profiles/{id}/results:
        get:
            description: |-
                Fetches pages of results from provider for the given id

                ```?order={field}``` orders on the passed field

                ```?page={page-number}``` Default page number is 0

                ```?pagesize={pagesize}``` Default pagesize is 10

                ```?search={result_name|mesh|url}``` If search is non empty then a greedy search is performed
            operationId: idGETProfileResults
            parameters:
                - description: id for a specific
                  format: uuid
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/performanceResultsResponseWrapper'
            summary: Handle GET request for results of a profile
            tags:
                - PerformanceAPI
...
produces:
    - application/json
responses:
...
    performanceResultsResponseWrapper:
        description: Returns all performance results
...
schemes:
    - http
security:
    - token:
        - '[]'
securityDefinitions:
    token:
        in: cookie
        name: token
        type: apiKey
swagger: "2.0"